fix: improve validate columns (#829)
* improve validate columns

* skipif

* skipif
